+++ genetic_algorithm/utils/sklearn_custom_transformers.py
@@ -1,24 +1,22 @@
-'''
-Contains custom transformers for scikit-learn
-'''
-
-# Imports ==========
+from typing import Optional
 
 import numpy as np
 import pandas as pd
 
 from sklearn.base import BaseEstimator, TransformerMixin
 
-# Create target mean encoder ==========
-
 class TargetMeanEncoder(BaseEstimator, TransformerMixin):
     
-    def __init__(self, prior_size=0, prior_frac=0):
+    def __init__(
+        self,
+        priorSize: Optional[int] = None, 
+        priorFrac: Optional[float] = None
+    ):
         
-        assert not (prior_size > 0 and prior_frac > 0)
+        assert not (priorSize > 0 and priorFrac > 0)
         
-        self.prior_size = prior_size
-        self.prior_frac = prior_frac
+        self.priorSize = priorSize
+        self.priorFrac = priorFrac
         
     def fit(self, X, y):
         
@@ -29,70 +27,73 @@ class TargetMeanEncoder(BaseEstimator, TransformerMixin):
         elif np.ndim(y) == 1:
             y = y.reshape(y.shape[0], 1)
 
-        data_arr = np.concatenate((X, y), axis=1)
+        dataArr = np.concatenate((X, y), axis=1)
         
-        grand_mean = np.mean(y)
+        grandMean = np.mean(y)
         
-        lvl_means  = {col: {} for col in range(X.shape[1])}
-        lvl_counts = {col: {} for col in range(X.shape[1])}
+        levelMeans = {j: {} for j in range(X.shape[1])}
+        levelCounts = {j: {} for j in range(X.shape[1])}
 
-        for col in lvl_means.keys():
-            for lvl in np.unique(X[:, col]):
+        for j in levelMeans.keys():
+            for g in np.unique(X[:, j]):
 
-                X_col_lvl = data_arr[:, col] == lvl
-                y_col     = data_arr.shape[1] - 1   
+                X_jg = dataArr[:, j] == g
+                y_j = dataArr.shape[1] - 1   
                 
-                X_col_lvl_y_numeric = data_arr[X_col_lvl, y_col].astype(float)
+                X_jg_y = data_arr[X_jg, y_j].astype(float)
                 
-                lvl_means[col][lvl]  = (
-                    np.mean(X_col_lvl_y_numeric) 
-                    if not np.isnan(np.mean(X_col_lvl_y_numeric))
+                lvl_means[j][g]  = (
+                    np.mean(X_jg_y) 
+                    if not np.isnan(np.mean(X_jg_y))
                     else grand_mean
                 )
                 
-                lvl_counts[col][lvl] = data_arr[X_col_lvl].shape[0]
+                lvl_counts[j][g] = dataArr[X_jg].shape[0]
                 
-        self.lvl_means = lvl_means
+        self.levelMeans = levelMeans
                 
-        if self.prior_size or self.prior_frac:
+        if self.prior_size > 0 or self.prior_frac > 0:
             
-            lvl_means_smoothed = {col: {} for col in range(X.shape[1])}
+            levelMeansSmoothed = {j: {} for j in range(X.shape[1])}
             
-            for col in lvl_means.keys():
-                for lvl in lvl_means[col].keys():
+            for j in levelMeans.keys():
+                for g in levelMeans[j].keys():
                     
-                    if self.prior_size:
-                        smooth_wt = self.prior_size / (self.prior_size + lvl_counts[col][lvl])
-                    elif self.prior_frac:
-                        prior_size = X.shape[0]*self.prior_frac
-                        smooth_wt  = prior_size / (prior_size + lvl_counts[col][lvl])
+                    if self.priorSize > 0:
+                        weightSmoothed = self.priorSize / (self.priorSize + levelCounts[j][g])
+                    elif self.priorFrac > 0:
+                        priorSize = X.shape[0]*self.priorFrac
+                        weightSmoothed = priorSize / (priorSize + levelCounts[j][g])
                         
-                    lvl_means_smoothed[col][lvl] = (
-                        (1 - smooth_wt)*lvl_means[col][lvl] + smooth_wt*grand_mean
-                        if not np.isnan((1 - smooth_wt)*lvl_means[col][lvl] + smooth_wt*grand_mean)
+                    levelMeansSmoothed_jg = (
+                        (1 - weightSmoothed)*levelMeans[j][g] + weightSmoothed*grandMean
+                    )                        
+                    levelMeansSmoothed[j][g] = (
+                        levelMeansSmoothed_jg
+                        if not np.isnan(levelMeansSmoothed_jg)
                         else grand_mean
                     )
                         
-            self.lvl_means_smoothed = lvl_means_smoothed
-            self.grand_mean = grand_mean
+            self.levelMeansSmoothed = levelMeansSmoothed
+            self.grandMean = grandMean
                 
         return self
         
-    def transform(self, X):
+    def transform(self, X: np.array) -> np.array:
         
         X = X.values if isinstance(X, pd.DataFrame) else X
-        X_transformed = np.empty_like(X).astype(float)
+        XTransformed = np.empty_like(X).astype(float)
         
-        for col in range(X_transformed.shape[1]):
+        for j in range(XTransformed.shape[1]):
             
-            if self.prior_size or self.prior_frac:
-                means_getter = np.vectorize(self.lvl_means_smoothed[col].get)
+            if self.priorSize > 0 or self.priorFrac > 0:
+                meansGetter = np.vectorize(self.levelMeansSmoothed[j].get)
             else:
-                means_getter = np.vectorize(self.lvl_means[col].get)
+                meansGetter = np.vectorize(self.levelmeans[j].get)
                 
-            X_transformed[:, col] = means_getter(X[:, col])
+            XTransformed[:, j] = meansGetter(X[:, j])
             
             # imputes for levels in validation set not seen in training set
-            X_transformed[np.isnan(X_transformed[:, col]), col] = self.grand_mean
+            XTransformed[np.isnan(XTransformed[:, j]), j] = self.grandMean
                 
-        return X_transformed
+        return XTransformed
